refactor(describe)!: extract ext-php-rs-introspection crate
