+++ Cargo.toml
@@ -26,6 +26,7 @@ smartstring = { version = "1", optional = true }
 indexmap = { version = "2", optional = true }
 inventory = "0.3"
 ext-php-rs-derive = { version = "=0.11.14", path = "./crates/macros" }
+ext-php-rs-introspection = { version = "0.1.0", path = "./crates/introspection" }
 
 [dev-dependencies]
 skeptic = "0.13"
@@ -62,7 +63,13 @@ runtime = ["ext-php-rs-bindgen/runtime"]
 static = ["ext-php-rs-bindgen/static"]
 
 [workspace]
-members = ["crates/macros", "crates/cli", "crates/php-build", "tests"]
+members = [
+    "crates/macros",
+    "crates/cli",
+    "crates/php-build",
+    "crates/introspection",
+    "tests",
+]
 
 [package.metadata.docs.rs]
 rustdoc-args = ["--cfg", "docs"]
