+++ crates/cli/src/ext.rs
@@ -1,10 +1,9 @@
 use std::path::PathBuf;
 
 use anyhow::{Context, Result};
-use ext_php_rs::describe::Description;
+use ext_php_rs_introspection::Description;
 use libloading::os::unix::{Library, RTLD_LAZY, RTLD_LOCAL, Symbol};
 
-#[allow(improper_ctypes_definitions)]
 pub struct Ext {
     // These need to be here to keep the libraries alive. The extension library needs to be alive
     // to access the describe function. Missing here is the lifetime on `Symbol<'a, fn() ->
